Cleanup Travis config

+++ .travis/travis-before-install.sh
@@ -1,60 +1,31 @@
 #!/usr/bin/env bash
 
-SWARM_VERSION="${SWARM_VERSION:-}"
-FAST_BUILD="${FAST_BUILD:-}"
-
-## fix coverity issue
-sudo apt-get install -y -q ca-certificates
-echo -n | openssl s_client -connect scan.coverity.com:443 | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' | sudo tee -a /etc/ssl/certs/ca-certificates.crt
-##
-
-if [ "$FAST_BUILD" == "true" ]; then
-    echo "Fast build, skipping docker installations."
-    exit 0
-fi
-
 set -exu
 
-sudo ip a ls
-sudo ip r ls
-sudo ss -antpl
+SWARM_VERSION="${SWARM_VERSION:-}"
+DOCKER_VERSION="${DOCKER_VERSION:-}"
 
 export HOST_PORT="2375"
-export SWARM_PORT="2377"
-export HOST_IP="$(ip a show dev eth0 | grep "inet\b" | awk '{print $2}' | cut -d/ -f1)"
-# because of swarm use docker-engine directly
-export PRE_DOCKER_HOST="$DOCKER_HOST"
-export DOCKER_HOST="tcp://127.0.0.1:${HOST_PORT}"
 
-
-docker info
-docker version
-
-sudo -E apt-get update
-sudo -E apt-get install -q -y wget
-sudo -E apt-get -q -y --purge remove docker-engine
-sudo -E apt-cache policy docker-engine
-
-./.travis/get-docker-com.sh
-
-sudo -E stop docker
-
-#mkdir "${HOME}/.cache" || :
-#pushd "${HOME}/.cache"
-# wget -N "https://apt.dockerproject.org/repo/pool/main/d/docker-engine/docker-engine_${DOCKER_VERSION}_amd64.deb"
-# sudo apt-get -f install
-# sudo dpkg -i "$(ls *${DOCKER_VERSION}*)"
-#popd
 rm -f "docker-java/src/test/resources/logback.xml"
-#rm -f "src/test/resources/travis-logback.xml"
 mv "docker-java/src/test/resources/travis-logback.xml" "docker-java/src/test/resources/logback-test.xml"
 
-# https://github.com/docker/docker/issues/18113
-sudo rm /var/lib/docker/network/files/local-kv.db
+if [[ -n $DOCKER_VERSION ]]; then
+    sudo -E apt-get update
+    sudo -E apt-get install -q -y wget
+    sudo -E apt-get -q -y --purge remove docker-engine
+    sudo -E apt-cache policy docker-engine
 
-sudo cat /etc/default/docker
+    curl -sSL https://get.docker.com/ | sh
 
-cat << EOF | sudo tee /etc/default/docker
+    sudo -E stop docker
+
+    # https://github.com/docker/docker/issues/18113
+    sudo rm /var/lib/docker/network/files/local-kv.db
+
+    sudo cat /etc/default/docker
+
+    cat << EOF | sudo tee /etc/default/docker
 DOCKER_OPTS="\
 --dns 8.8.8.8 \
 --dns 8.8.4.4 \
@@ -65,39 +36,30 @@ DOCKER_OPTS="\
 "
 EOF
 
-sudo cat /etc/default/docker
-sudo bash -c ':> /var/log/upstart/docker.log'
+    sudo cat /etc/default/docker
+    sudo bash -c ':> /var/log/upstart/docker.log'
 
-date
-sudo -E start docker
+    sudo -E start docker
 
-tries=20
-sleep=5
-for i in $(seq 1 $tries); do
-    if sudo grep "API listen on" /var/log/upstart/docker.log ; then
-        echo "Docker started. Delay $(($i * $sleep))"
-        break
-    elif [[ $i -ge $tries ]]; then
-        echo "Docker didn't start. Exiting!"
-        sudo cat /var/log/upstart/docker.log
-        exit 1
-    else
-        echo "Docker didn't start, sleeping for 5 secs..."
-        sleep $sleep
-    fi
-done
+    tries=20
+    sleep=5
+    for i in $(seq 1 $tries); do
+        if sudo grep "API listen on" /var/log/upstart/docker.log ; then
+            echo "Docker started. Delay $(($i * $sleep))"
+            break
+        elif [[ $i -ge $tries ]]; then
+            echo "Docker didn't start. Exiting!"
+            sudo cat /var/log/upstart/docker.log
+            exit 1
+        else
+            echo "Docker didn't start, sleeping for 5 secs..."
+            sleep $sleep
+        fi
+    done
 
+    set +u
 
-sudo ss -antpl
-
-curl -V
-
-docker version || sudo cat /var/log/upstart/docker.log
-docker info
-
-set +u
-
-cat <<EOF > "${HOME}/.docker-java.properties"
+    cat <<EOF > "${HOME}/.docker-java.properties"
 registry.username=${registry_username}
 registry.password=${registry_password}
 registry.email=${registry_email}
@@ -105,35 +67,26 @@ registry.url=https://index.docker.io/v1/
 
 EOF
 
+fi
+
+docker version || sudo cat /var/log/upstart/docker.log
+docker info
+
 if [[ -n $SWARM_VERSION ]]; then
-#    export SWARM_PORT="${PRE_DOCKER_HOST##*:}"
+    export SWARM_PORT="2377"
+    export HOST_IP="$(ip a show dev eth0 | grep "inet\b" | awk '{print $2}' | cut -d/ -f1)"
+    # because of swarm use docker-engine directly
+    export PRE_DOCKER_HOST="$DOCKER_HOST"
+    export DOCKER_HOST="tcp://127.0.0.1:${HOST_PORT}"
 
     docker pull swarm
-
-#    # kv store https://docs.docker.com/v1.11/engine/userguide/networking/get-started-overlay/
-#    docker run -d \
-#        -p "8500:8500" \
-#        -h "consul" \
-#        --name=consul \
-#        progrium/consul -server -bootstrap
-#
-#    sleep 5
-
-#    SWARM_TOKEN=$(docker run swarm c)
-
-#    docker run \
-#        -d \
-#        --name=swarm_manager \
-#        -p ${SWARM_PORT}:2375 \
-#        "swarm:${SWARM_VERSION}" \
-#        manage token://${SWARM_TOKEN}
 
     docker run \
         -d \
         -p ${SWARM_PORT}:2375 \
         --name=swarm_manager \
-        swarm manage --engine-refresh-min-interval "3s" --engine-refresh-max-interval "6s" "nodes://${HOST_IP}:${HOST_PORT}"
-#        swarm manage --engine-refresh-min-interval "3s" --engine-refresh-max-interval "6s" "consul://${HOST_IP}:8500"
+        "swarm:${SWARM_VERSION}" \
+        manage --engine-refresh-min-interval "3s" --engine-refresh-max-interval "6s" "nodes://${HOST_IP}:${HOST_PORT}"
 
     # join engine to swarm
     docker run \
@@ -141,19 +94,17 @@ if [[ -n $SWARM_VERSION ]]; then
         "--name=swarm_join" \
         "swarm:${SWARM_VERSION}" \
         join --advertise="${HOST_IP}:${HOST_PORT}" --delay="0s" --heartbeat "5s" "nodes://${HOST_IP}:${HOST_PORT}"
-#        join --advertise="${HOST_IP}:${HOST_PORT}" --delay="0s" --heartbeat "5s" "token://${SWARM_TOKEN}"
 
     docker run --rm \
-        "swarm:${SWARM_VERSION}" list "nodes://${HOST_IP}:${HOST_PORT}"
+        "swarm:${SWARM_VERSION}" \
+        list "nodes://${HOST_IP}:${HOST_PORT}"
 
     docker ps -a
-    sudo ss -antpl
 
     sleep 30
 
     docker logs swarm_join
     docker logs swarm_manager
-#    docker logs consul
 
     # switch to swarm connection
     DOCKER_HOST="$PRE_DOCKER_HOST"
